Added linear interpolation for `--echang` option.
Any float between 60.0 and 70.0 can be entered for the `--echang`
option, instead of the previous 0.5 step.
The values are taken using linear interpolation for the lower and
upper closest neighbors.

It is crudely implemented in the code, and some cleaning up could
be done (repeated look-ups, complicated handling of the CodeV
data tables, etc.).